Question 4 completed with gb_calc_size
