+++ report/tests/test_filesize.py
@@ -7,27 +7,19 @@ from report.text import Text
 
 class TestFileSize(unittest.TestCase):
 
-    def calc_gb_size(self, unit: str, size: int) -> float:
-        if unit == 'GB':
-            return round(size * 1024, 2)
-        else:
-            return round(size, 2)
-
-    def test_answer_large_gb_file(self):
-
+    def test_answer_large_mb_file(self):
         test_answer: Dict = {
-            'SIZE': 2,
-            'UNIT': 'GB'
+            'SIZE': 200,
+            'UNIT': 'MB'
         }
-
-        gb_to_mb: float = self.calc_gb_size(
-            unit=test_answer['UNIT'],
-            size=test_answer['SIZE']
-        )
 
         choices = Choices(answer={'FSIZE': test_answer})
         choices.filesize()
         text = Text()
+        gb_to_mb: float = text.calc_gb_size(
+            unit=test_answer['UNIT'],
+            size=test_answer['SIZE']
+        )
 
         self.assertEqual(
             choices.result_text,
@@ -35,21 +27,18 @@ class TestFileSize(unittest.TestCase):
             text.large_size_in_megabytes(gb_size=gb_to_mb)
         )
 
-    def test_answer_large_mb_file(self):
-
+    def test_answer_large_gb_file(self):
         test_answer: Dict = {
-            'SIZE': 200,
-            'UNIT': 'MB'
+            'SIZE': 2,
+            'UNIT': 'GB'
         }
-
-        gb_to_mb: float = self.calc_gb_size(
-            unit=test_answer['UNIT'],
-            size=test_answer['SIZE']
-        )
-
         choices = Choices(answer={'FSIZE': test_answer})
         choices.filesize()
         text = Text()
+        gb_to_mb: float = text.calc_gb_size(
+            unit=test_answer['UNIT'],
+            size=test_answer['SIZE']
+        )
 
         self.assertEqual(
             choices.result_text,
@@ -62,14 +51,14 @@ class TestFileSize(unittest.TestCase):
             'SIZE': 1,
             'UNIT': 'GB'
         }
-        gb_to_mb: float = self.calc_gb_size(
+        choices = Choices(answer={'FSIZE': test_answer})
+        choices.filesize()
+        text = Text()
+        gb_to_mb: float = text.calc_gb_size(
             unit=test_answer['UNIT'],
             size=test_answer['SIZE']
         )
 
-        choices = Choices(answer={'FSIZE': test_answer})
-        choices.filesize()
-        text = Text()
         self.assertEqual(
             choices.result_text,
             '\n' +
@@ -81,14 +70,14 @@ class TestFileSize(unittest.TestCase):
             'SIZE': 49,
             'UNIT': 'MB'
         }
-        gb_to_mb: float = self.calc_gb_size(
-            unit=test_answer['UNIT'],
-            size=test_answer['SIZE']
-        )
 
         choices = Choices(answer={'FSIZE': test_answer})
         choices.filesize()
         text = Text()
+        gb_to_mb: float = text.calc_gb_size(
+            unit=test_answer['UNIT'],
+            size=test_answer['SIZE']
+        )
         self.assertEqual(
             choices.result_text,
             '\n' +
